chore: add new shared-helpers package
BREAKING CHANGE: moves form keys out of ui-components
