+++ shared-helpers/.jest/setup-tests.js
@@ -1,0 +1,1 @@
+// Future home of additional Jest config
