+++ shared-helpers/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "@bloom-housing/shared-helpers",
+  "version": "1.0.5",
+  "description": "Shared helpers for Bloom affordable housing system",
+  "homepage": "https://github.com/bloom-housing/bloom/tree/master/shared-helpers",
+  "main": "index.js",
+  "license": "Apache-2.0",
+  "private": false,
+  "scripts": {
+    "test": "jest -w 1",
+    "test:coverage": "jest -w 1 --coverage --watchAll=false",
+    "prettier": "prettier --write \"**/*.ts\""
+  },
+  "devDependencies": {
+    "jest": "^26.5.3",
+    "typescript": "^3.9.7"
+  }
+}
