+++ shared-helpers/src/formKeys.ts
@@ -1,0 +1,180 @@
+export const stateKeys = [
+  "",
+  "AL",
+  "AK",
+  "AZ",
+  "AR",
+  "CA",
+  "CO",
+  "CT",
+  "DE",
+  "DC",
+  "FL",
+  "GA",
+  "HI",
+  "ID",
+  "IL",
+  "IN",
+  "IA",
+  "KS",
+  "KY",
+  "LA",
+  "ME",
+  "MD",
+  "MA",
+  "MI",
+  "MN",
+  "MS",
+  "MO",
+  "MT",
+  "NE",
+  "NV",
+  "NH",
+  "NJ",
+  "NM",
+  "NY",
+  "NC",
+  "ND",
+  "OH",
+  "OK",
+  "OR",
+  "PA",
+  "RI",
+  "SC",
+  "SD",
+  "TN",
+  "TX",
+  "UT",
+  "VT",
+  "VA",
+  "WA",
+  "WV",
+  "WI",
+  "WY",
+]
+
+export const contactPreferencesKeys = [
+  {
+    id: "email",
+  },
+  {
+    id: "phone",
+  },
+  {
+    id: "letter",
+  },
+  {
+    id: "text",
+  },
+]
+
+export const relationshipKeys = [
+  "",
+  "spouse",
+  "registeredDomesticPartner",
+  "parent",
+  "child",
+  "sibling",
+  "cousin",
+  "aunt",
+  "uncle",
+  "nephew",
+  "niece",
+  "grandparent",
+  "greatGrandparent",
+  "inLaw",
+  "friend",
+  "other",
+]
+
+export const altContactRelationshipKeys = [
+  "familyMember",
+  "friend",
+  "caseManager",
+  "other",
+  "noContact",
+]
+
+export const ethnicityKeys = ["hispanicLatino", "notHispanicLatino"]
+
+export const raceKeys = [
+  "americanIndianAlaskanNative",
+  "asian",
+  "blackAfricanAmerican",
+  "nativeHawaiianOtherPacificIslander",
+  "white",
+  "americanIndianAlaskanNativeAndBlackAfricanAmerican",
+  "americanIndianAlaskanNativeAndWhite",
+  "asianAndWhite",
+  "blackAfricanAmericanAndWhite",
+  "otherMutliracial",
+]
+
+export const genderKeys = [
+  "female",
+  "male",
+  "genderqueerGenderNon-Binary",
+  "transFemale",
+  "transMale",
+  "notListed",
+]
+
+export const sexualOrientation = [
+  "bisexual",
+  "gayLesbianSameGenderLoving",
+  "questioningUnsure",
+  "straightHeterosexual",
+  "notListed",
+]
+
+export const howDidYouHear = [
+  {
+    id: "alamedaCountyHCDWebsite",
+  },
+  {
+    id: "developerWebsite",
+  },
+  {
+    id: "flyer",
+  },
+  {
+    id: "emailAlert",
+  },
+  {
+    id: "friend",
+  },
+  {
+    id: "housingCounselor",
+  },
+  {
+    id: "radioAd",
+  },
+  {
+    id: "busAd",
+  },
+  {
+    id: "other",
+  },
+]
+
+export const phoneNumberKeys = ["work", "home", "cell"]
+
+export const preferredUnit = [
+  {
+    id: "studio",
+  },
+  {
+    id: "oneBedroom",
+  },
+  {
+    id: "twoBedroom",
+  },
+  {
+    id: "threeBedroom",
+  },
+  {
+    id: "moreThanThreeBedroom",
+  },
+]
+
+export const bedroomKeys = ["studio", "oneBdrm", "twoBdrm", "threeBdrm"]
