+++ sites/public/pages/applications/contact/alternate-contact-contact.tsx
@@ -16,7 +16,7 @@ import FormsLayout from "../../../layouts/forms"
 import { useForm } from "react-hook-form"
 import { Select } from "@bloom-housing/ui-components/src/forms/Select"
 import { PhoneField } from "@bloom-housing/ui-components/src/forms/PhoneField"
-import { stateKeys } from "@bloom-housing/ui-components/src/helpers/formOptions"
+import { stateKeys } from "@bloom-housing/shared-helpers"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
 import { useFormConductor } from "../../../lib/hooks"
